- Fix 64bit bug that bites on HP-UX 11.x
(Logical change 1.49)


git-svn-id: https://svn.code.sf.net/p/lmbench/code/trunk@633 9cc621ae-7264-4df2-8f7e-c588375cf588

+++ lmbench2/src/lmdd.c
@@ -664,7 +664,7 @@ done(void)
 	if (hash || poff) {
 		fprintf(stderr, "\n");
 	}
-	if ((int)Label != -1) {
+	if ((long)Label != -1) {
 		fprintf(stderr, "%s", Label);
 	}
 	int_count <<= 2;
